Remove access to global state and deprecated APIs
The main change is to remove any access to `Play.maybeApplication` which
Play does not set by default anymore. See playframework/playframework#8462.

This implies changing tests that also access the global state when using the
`Action` object. So now all the tests are instead using `appRoutes` instead of
`routes` since the later does not provides a way to access an ActionBuilder.

Finally, some minor deprecated calls were replaced with the new APIs.